requirements.txt
terminaltables==3.1.10
python-dotenv==0.21.1
requests==2.28.2